requirements.txt
fastapi
uvicorn
sentence-transformers
gensim
stanza
networkx
python-dotenv
pydantic
llama-index
openai
transformers
requests
pydantic-settings
httpx[socks]